# perf(ingest): speed up MCPW.validate()