In my local tests with an MCPW with a dataset schema of 3000 columns, this reduced runtime from ~676 ms to ~189 ms, which is a 3-4x improvement. It seems to scale linearly with the size of the MCPW object, but I didn't test that extensively.

Note that almost 200ms is still pretty bad for basic data validation. With the changes from https://github.com/acryldata/avro_gen/pull/18, I was able to get it down to ~130 ms. Not sure how much further it is worth pushing it.


## Checklist

- [ ] The PR conforms to DataHub's [Contributing Guideline](https://github.com/datahub-project/datahub/blob/master/docs/CONTRIBUTING.md) (particularly [Commit Message Format](https://github.com/datahub-project/datahub/blob/master/docs/CONTRIBUTING.md#commit-message-format))
- [ ] Links to related issues (if applicable)
- [ ] Tests for the changes have been added/updated (if applicable)
- [ ] Docs related to the changes have been added/updated (if applicable). If a new feature has been added a Usage Guide has been added for the same.
- [ ] For any breaking change/potential downtime/deprecation/big changes an entry has been made in [Updating DataHub](https://github.com/datahub-project/datahub/blob/master/docs/how/updating-datahub.md)